Moved download related files to dedicated folder
