+++ libmamba/include/mamba/core/download_progress_bar.hpp
@@ -12,8 +12,8 @@
 #include <vector>
 
 #include "mamba/core/context.hpp"
-#include "mamba/core/download.hpp"
 #include "mamba/core/package_fetcher.hpp"
+#include "mamba/download/downloader.hpp"
 
 namespace mamba
 {
